[shape_poly] Fixes for approx_top_k when `aggregated_to_topk=True`
When `aggregate_to_topk=True` (the default) the output reduction
dimension size is `k`, and we do not need to invoke `ApproxtopKReductionOutputSize`.

Add a set of test cases for shape polymorphism for approx_top_k.

The case when `aggregate_to_topk=True` and `k` is symbolic will
be fixed separately.

The case when `aggregate_to_topk=False` raises a clearer NotImplementedError.