Wrap wgmma.fence in llvm.inline_asm to constrain LLVM scheduling
wgmma.fence.aligned is a weird PTX instruction in that it is one of the
few (if not the only one?) that disallows sinking ALU ops on registers
below it. But, LLVM assumes that all operations on registers are pure
and will often happily sink them below this instruction. By wrapping
the fence in an inline assembly block that simply copies over the
registers, we can force LLVM to construct the registers before the fence.
And ptxas should be able to eliminate the unnecessary register copies.

PiperOrigin-RevId: 639011288