[Mosaic GPU] Skip tests for Python versions < 3.10
We use the match statement in Mosaic GPU, making it incompatible with Python 3.9.
Since we're dropping support for 3.9 in about a month, it's better to skip the
OSS tests for now (especially that we can't test with H100s) than to rewrite the
code.

Also add a check for having at least an sm90 GPU.

PiperOrigin-RevId: 639733266