Make notebook execution on Read the Docs explicitly opt-in
This is an initial attempt to speed up the Read the Docs build time so
that we don't end up with a backed up build queue. The idea is to
(eventually) only execute the notebooks if explicitly requested
(potentially by adding a label or a tag in the commit message). This
means that the rendered previews won't be complete, but we should still
be able to catch notebook bugs on GitHub Actions.