Update Windows Actions workflow name
When skipped, the build matrix for the Windows workflow on GitHub
Actions doesn't get populated, so the run name ends up as
`${{ matrix.os }} CI build`. It doesn't seem necessary to include the
matrix value in the name since there is only one build anyways. This PR
replaces the name with a constant.