[Mosaic GPU] Call mbarrier.try_wait only once mbarrier.test_wait fails
The llvm.expect intrinsic puts the loop at the end of the program, allowing
the whole barrier to be compiled to a test_wait + predicated branch that is
immediately followed by the continuation. This seems to make the happy path
a little faster which can help reduce the barrier overhead for compute-bound
kernels.

PiperOrigin-RevId: 645007019