Apply ruff linting to all docstrings.
I used a combination of ruff's `--fix` option and manual editing to get
all the doctrings to the be a little more stylistically consistent. This
is a huge PR, but the only changes are to docstrings, so maybe that's
ok. I'd also be happy to break this up into smaller chunks if that would
be better for reviewing.

I've left in the ruff config that I was using, but I don't feel strongly
about actually including these checks as part of our pre-commit; I'd be
happy to revert the changes to `pyproject.toml`!