[pallas] Simplify the generation of the hlo.CustomCall
We use the `mlir.custom_call` library that most other
JAX lowering rules use.

We remove the use of `hlo.TupleType` because custom calls
support multiple results. Tuples are not supported
by the `mlir.custom_call` and they complicate the logic
for input_output_aliasing and for the dynamically-shaped
results.

PiperOrigin-RevId: 646462934