Add installation instructions back to documentation homepage.
These instructions were removed in #17744, but it has been suggested
that it would be useful to have easy access to this info directly from
the docs homepage. This results in some duplication of the instructions
(here, README, and `installation.md`), but it doesn't seem too onerous
to keep these synchronized.