[pallas] Make pallas_test run on CPU and TPU also.
pallas_test was only running on GPU, but it is useful to run this test on all platform, in both interpret mode and the native mode. Added `skipTest` and `TODO` for the tests that fail, and in some cases configured numerical comparison tolerances.

All tests now have a "Interpreter" version, e.g., for `CallTest` we also define a `CallInterpreterTest` that runs the same tests but in interpreter
mode. This was not done systematically before, and in some cases the
interpreter test was missing, or was empty.

Some of the tests in pallas_test perhaps make sense only for GPU. I will
split them out in a separate CL.

PiperOrigin-RevId: 648619580