Fix host compute annotations in the presence of copies.
If an explicit copy operation was present in a function offload to the host, the adding of compute annotations during lowering failed since the operation lowered to a block argument and the "owner" of the block argument is the block - not an operation. Fixed by filtering out this case when adding compute annotations.

PiperOrigin-RevId: 653151844