[shard_map] fix psum rewrite rule's pbroadcast logic
We want to allow `psum(x, axes)` regardless of how `x` is replicated. That
means when we rewrite it into the stricter `psum2`, which can only sum over
non-replicated axes, we need to insert a pbroadcast like this:

```
psum(x, axes) == psum2(pbroadcast(x, axes & input_replicated_axes), axes)
```

In words, we need to first `pbroadcast` over all those axes we're about to sum
over but that the input is already replicated over.

We write it as a comprehension over mesh.axis_names, rather than just that set
intersection, just to ensure deterministic ordering, since Python set
operations are not guaranteed to be deterministic. There are other places in
the file where we don't ensure deterministic ordering; someday I'll come back
and fix those.

fixes #19175