Add support for using dynamic shapes with GPU LU decomposition.
After switching to the FFI for this custom call, it is straightforward to support dynamic shapes in all axes (including the non batch dimensions). This does come with one potential performance caveat: when the non batch dimensions are statically known at lowering time and equal (`m == n`), we lower to the cuBLAS "batched" implementation which is significantly faster. When `m` and/or `n` are dynamic we always lower to the cuSolver version.

It would actually be straightforward to defer this decision (cuBLAS vs. cuSolver) into the kernel, and I found that that didn't have any obvious performance implications, but I haven't figured out how to do the same thing for the QR decomposition because the two versions return different outputs. I'll revisit this detail in a follow up.

With this change, the GPU and CPU lowering rules are now nearly identical. In another follow up I will consolidate those implementations.

PiperOrigin-RevId: 659894314