Port jaxlib's triton_call target to the FFI.
This is a minimal port of the current triton custom call to the new typed FFI. I say "minimal" here because this implementation reuses almost all of the existing structure and doesn't really take advantage of the features that the FFI provides.

I have only just started digging into the guts of this kernel to identify places where we could improve the API using the FFI, but it might make sense to do those experiments closer to jax-triton where it would be easier to experiment with the performance implications.

PiperOrigin-RevId: 667665618