Added shape and dtype to AbstractValue
We assume that an `AbstractValue` is a `ShapedArray` in lots of places, and
currently most of them generate static type errors, because an `AbstractValue`
has neither `shape` nor `dtype` (unliked `ShapedArray`).

I decided to use `tuple[Any, ...]` instead of `tuple[int, ...]` for the `shape`
because dimension are not `int`s in the presence of dynamic shapes.

With this

     python -m pyright jax/_src

is down to

     2063 errors, 19 warnings, 0 informations