Unconditionally lower jnp.dot to lax.dot_general.
https://github.com/google/jax/pull/16721 added a condition to lower
calls to `jnp.dot` with scalar inputs to `lax.mul` instead of
`lax.dot_general`. AFAICT, https://github.com/google/jax/pull/16826
fixed the issue that this was solving, so this condition should no
longer be necessary. Removing this condition simplifies the addition of
new arguments to `dot` and `dot_general`, including the `algorithm`
parameter that I am currently working on in
https://github.com/google/jax/pull/23574, so now seemed like a good time
to remove it!