[host_callback] Adjust the legacy compatibility mode for call_with_device
The jax.experimental.host_callback module is deprecated and will be removed.
See https://github.com/google/jax/issues/20385.

The next step in the removal is to flip the flag `--jax_host_callback_legacy` to `False`, meaning that the host callback APIs will be implemented by way of `io_callback`. In this mode, the `call_with_device` and `tap_with_device` cannot be supported.

We add a temporary mechanism to be used internally for a couple of projects that still use `call_with_device`: allow the use but force the device to be the first device. This mechanism is not used externally, and attempting to use `call_with_device` will continue to raise an error.

PiperOrigin-RevId: 676720231