[Mosaic TPU] Improve vector layout inference for vector.shape_cast
This both simplifies and strengthens the pass. The loop over candidate result implicit
dims was not useful beyond the initial no-op version. I don't think that there's any
implicit dim that would in fact satisfy the checks in a way that the current implementation
doesn't.

What's more important, however, is that the previous implementation never tried to alter
the implicit dim of the input layout, which caused unnecessary failures when inputs came
in with an implicit_dim. Now the rules are more explicit and will proactively adjust the
input layout to make the shape_cast realizable.

PiperOrigin-RevId: 676784915